feat(opponent): :sparkles: add random bot
