+++ ChessApp/Features/Chess/ChessState.cs
@@ -11,9 +11,7 @@ public partial class ChessState : State<ChessState>
 
     public Chessboard Board { get; private set; }
 
-    public Player WhitePlayer { get; set; }
-
-    public Player BlackPlayer { get; set; }
+    
 
     public override void Initialize()
     {
@@ -23,17 +21,6 @@ public partial class ChessState : State<ChessState>
         MovingPositon = new Position('0', 0);
     }
 
-    public void StartGame(Player White = Player.This, Player Black = Player.This)
-    {
-        WhitePlayer = White;
-        BlackPlayer = Black;
-        Board.GameState = GameState.Playing;
-    }
+    
 }
 
-public enum Player
-{
-    This,
-    Bot,
-    OnlinePlayer
-}
