+++ ChessApp/Pages/Index.razor
@@ -19,6 +19,6 @@
     protected override void OnInitialized()
     {
         chessState.Initialize();
-        chessState.StartGame();
+        chessState.Board.StartGame();
     }
 }
