+++ ChessApp/Pages/PlayBot.razor
@@ -19,5 +19,6 @@
     protected override void OnInitialized()
     {
         chessState.Initialize();
+        chessState.Board.StartGame(Black: Player.RandomBot);
     }
 }}
